pyproject.toml
[tool.poetry]
name = "elysian-assignment"
version = "0.1.0"
description = "Description of your project"
authors = ["Your Name <youremail@example.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.100.0"  # Updated to a version known to support Pydantic 2.x
uvicorn = "^0.18.0"  # You can update uvicorn as well to a recent compatible version
sqlalchemy = "<2.0"
alembic = "^1.7"
pydantic = "^2.7"
pydantic-settings = "^2.0"
pymysql = "^1.0.2"
python-dotenv = "^1.0.1"
openai = "^1.34.0"
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"
bandit = "^1.7.9"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
pytest-asyncio = "^0.19.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"